Added more tests + cleaned up test suite
